scripts/finetuning/requirements.txt
pdfplumber==0.11.0
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.0
tqdm==4.66.1
python-dotenv==1.0.1
lxml==5.1.0
google-generativeai==0.8.3
openai==1.55.0
mistralai==1.2.1
